requirements.txt
celery==4.3.0
Django==2.0.7
django-cors-headers==3.0.2
django-filter==2.2.0
django-redis==4.10.0
django-redis-sessions==0.5.6
djangorestframework==3.10.2
Pillow==6.1.0
PyMySQL==0.9.3
redis==3.3.6
